backend/requirements.txt
# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart

# Azure OpenAI
openai>=1.12.0
httpx==0.27.0

# Vector store
faiss-cpu==1.7.4
numpy>=1.24.0,<2.0  # Pin to <2.0 for faiss-cpu compatibility

# PDF processing
pypdf>=4.0.0

# Text splitting
langchain-text-splitters>=0.0.7

# Environment
python-dotenv>=1.0.0
